+++ backend/src/main/java/com/kostiago/backend/entities/Product.java
@@ -25,7 +25,7 @@ public class Product implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -60,7 +60,7 @@ public class Product implements Serializable {
     public Product() {
     }
 
-    public Product(Integer id, String name, String shortDescription, String description, Double price, Double sale,
+    public Product(Long id, String name, String shortDescription, String description, Double price, Double sale,
             Instant createDate, Instant updateDate) {
         this.id = id;
         this.name = name;
@@ -73,11 +73,11 @@ public class Product implements Serializable {
 
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
